requirements.txt
# Core
pandas>=2.0.3
numpy>=1.24.3
python-dateutil==2.8.2
pytz==2023.3
python-dotenv

# HTTP / API
requests==2.31.0

# Config
python-dotenv==1.0.0

# Market data
yfinance>=0.2.40

# Scheduling (optional for local scheduler)
APScheduler==3.10.4

# Telegram
python-telegram-bot==20.3

# Timezone helpers
tzdata==2023.3
google-cloud-firestore
fyers-apiv3>=3.0.0
google-cloud-secret-manager>=2.16.0

# Machine Learning
lightgbm>=4.1.0
scikit-learn>=1.3.0
google-cloud-storage>=2.10.0  # For GCS model storage
google-cloud-aiplatform>=1.38.0  # For Vertex AI Gemini
